feat(backend): implement authenticate and authorize to label mutation routes
Implemented authenticate and GlobalRole.Admin for all label mutation routes

+++ server/src/router/labelRoutes.ts
@@ -6,6 +6,9 @@ import {
   updateLabel,
   deleteLabel,
 } from '../controllers/labelController';
+import { authenticate } from '../middleware/authenticate';
+import { authorize } from '../middleware/authorize';
+import { GlobalRole } from '@prisma/client';
 
 const router = Router();
 
@@ -17,6 +20,8 @@ router.get('/labels', async (req: Request, res: Response): Promise<void> => {
 // Create new Label
 router.post(
   '/labels',
+  authenticate,
+  authorize(GlobalRole.ADMIN),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await createNewLabel(req, res, next, prisma);
   }
@@ -25,6 +30,8 @@ router.post(
 // Update Label
 router.patch(
   '/labels/:labelId',
+  authenticate,
+  authorize(GlobalRole.ADMIN),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await updateLabel(req, res, next, prisma);
   }
@@ -33,6 +40,8 @@ router.patch(
 // Delete Label
 router.delete(
   '/labels/:labelId',
+  authenticate,
+  authorize(GlobalRole.ADMIN),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await deleteLabel(req, res, next, prisma);
   }
